+++ src/Candoumbe.Pipelines/Components/IMutationTest.cs
@@ -2,12 +2,15 @@
 using System.Collections.Generic;
 using System.Linq;
 using Candoumbe.Pipelines.Components.Workflows;
+using Candoumbe.Pipelines.Tools;
+using JetBrains.Annotations;
 using Nuke.Common;
 using Nuke.Common.IO;
 using Nuke.Common.ProjectModel;
 using Nuke.Common.Tooling;
-using static Nuke.Common.Tools.DotNet.DotNetTasks;
 using static Serilog.Log;
+using static Candoumbe.Pipelines.Tools.StrykerTasks;
+
 namespace Candoumbe.Pipelines.Components;
 
 /// <summary>
@@ -17,7 +20,6 @@ namespace Candoumbe.Pipelines.Components;
 /// <see cref="MutationTests"/> target required <see href="https://www.nuget.org/packages/dotnet-stryker">Stryker dotnet tool</see> to be referenced in order to run.<br />
 /// Simply adds <c>&lt;PackageReference Include="dotnet-stryker" Version="" ExcludeAssets="all" &gt;</c> to your build project file
 /// </remarks>
-[NuGetPackageRequirement("dotnet-stryker")]
 public interface IMutationTest : IHaveTests
 {
     /// <summary>
@@ -95,48 +97,42 @@ public interface IMutationTest : IHaveTests
 
                 Verbose("{ProjectName} will run mutation tests for the following frameworks : {@Frameworks}", sourceProject.Name, sourceProject.GetTargetFrameworks());
 
-                Arguments args = new();
-                args = args.Apply(StrykerArgumentsSettingsBase)
-                           .Apply(StrykerArgumentsSettings);
+                StrykerSettings strykerArgs = new();
+                strykerArgs = StrykerArgumentsSettingsBase.Invoke(strykerArgs);
+                strykerArgs = StrykerArgumentsSettings.Invoke(strykerArgs);
 
-                args.Add("--target-framework {value}", framework)
-                    .Add("--output {value}", MutationTestResultDirectory / sourceProject.Name / framework);
-
-                Arguments strykerArgs = new();
-                strykerArgs = strykerArgs.Add("stryker");
-
-                strykerArgs = strykerArgs.Concatenate(args);
-
-                strykerArgs = strykerArgs.Add("--project {value}", $"{sourceProject.Name}{sourceProject.Path.Extension}");
+                strykerArgs = strykerArgs.SetTargetFramework(framework);
+                strykerArgs = strykerArgs.SetOutput(MutationTestResultDirectory / sourceProject.Name / framework);
+                strykerArgs = strykerArgs.SetProject($"{sourceProject.Name}{sourceProject.Path.Extension}");
 
                 if (configFile is not null)
                 {
-                    strykerArgs = strykerArgs.Add("--config-file {value}", configFile);
+                    strykerArgs = strykerArgs.SetConfigFile(configFile);
                 }
 
-                strykerArgs = testsProjects.Aggregate(strykerArgs, (current, project) => current.Add("--test-project {value}", project.Path));
+                strykerArgs = strykerArgs.SetTestProjects([.. testsProjects.Select(project => project.Path)]);
 
                 switch (this)
                 {
                     case IGitFlow { GitRepository: { } gitflowRepository } gitFlow:
                         {
-                            strykerArgs = strykerArgs.Add("--version {value}", gitflowRepository.Commit ?? gitflowRepository.Branch);
+                            strykerArgs = strykerArgs.SetProjectInfoVersion($"{gitflowRepository.Commit ?? gitflowRepository.Branch}");
                             switch (gitflowRepository.Branch)
                             {
                                 case { } branchName when string.Equals(branchName, IHaveDevelopBranch.DevelopBranchName, StringComparison.InvariantCultureIgnoreCase):
                                     {
                                         // we are in git flow, so we can compare develop with main branch
-                                        strykerArgs = strykerArgs.Add("--with-baseline:{value}", IHaveMainBranch.MainBranchName);
+                                        strykerArgs = strykerArgs.SetWithBaseline(IHaveMainBranch.MainBranchName);
                                     }
                                     break;
                                 case { } branchName when branchName.Like($"{gitFlow.FeatureBranchPrefix}/*", true):
                                     {
-                                        strykerArgs = strykerArgs.Add("--with-baseline:{value}", gitFlow.FeatureBranchSourceName);
+                                        strykerArgs = strykerArgs.SetWithBaseline(gitFlow.FeatureBranchSourceName);
                                     }
                                     break;
                                 case { } branchName when branchName.Like($"{gitFlow.ColdfixBranchPrefix}/*", true):
                                     {
-                                        strykerArgs = strykerArgs.Add("--with-baseline:{value}", gitFlow.ColdfixBranchSourceName);
+                                        strykerArgs = strykerArgs.SetWithBaseline(gitFlow.ColdfixBranchSourceName);
                                     }
                                     break;
                                 default:
@@ -144,7 +140,7 @@ public interface IMutationTest : IHaveTests
                                         // By default, we try to compare the current commit with the previous one (if any).
                                         if (gitflowRepository.Head is { } head)
                                         {
-                                            strykerArgs = strykerArgs.Add("--with-baseline:{value}", head);
+                                            strykerArgs = strykerArgs.SetWithBaseline(head);
                                         }
                                     }
                                     break;
@@ -155,10 +151,10 @@ public interface IMutationTest : IHaveTests
 
                     case IGitHubFlow { GitRepository: { } githubFlowRepository }:
                         {
-                            strykerArgs = strykerArgs.Add("--version {value}", githubFlowRepository.Commit ?? githubFlowRepository.Branch);
+                            strykerArgs = strykerArgs.SetProjectInfoVersion(githubFlowRepository.Commit ?? githubFlowRepository.Branch);
                             if (githubFlowRepository.Branch is { Length: > 0 } branchName && !string.Equals(branchName, IHaveMainBranch.MainBranchName, StringComparison.InvariantCultureIgnoreCase))
                             {
-                                strykerArgs = strykerArgs.Add("--with-baseline:{0}", IHaveMainBranch.MainBranchName);
+                                strykerArgs = strykerArgs.SetWithBaseline(IHaveMainBranch.MainBranchName);
                             }
 
                             break;
@@ -170,39 +166,36 @@ public interface IMutationTest : IHaveTests
                         {
                             strykerArgs = this switch
                             {
-                                IHaveGitVersion gitVersion => strykerArgs.Add("--version {value}", gitVersion.MajorMinorPatchVersion),
-                                IHaveGitRepository gitRepository => strykerArgs.Add("--version {value}", gitRepository.GitRepository?.Commit ?? gitRepository.GitRepository?.Branch),
+                                IHaveGitVersion gitVersion => strykerArgs.SetProjectInfoVersion(gitVersion.MajorMinorPatchVersion),
+                                IHaveGitRepository gitRepository => strykerArgs.SetProjectInfoVersion(gitRepository.GitRepository?.Commit ?? gitRepository.GitRepository?.Branch),
                                 _ => strykerArgs
                             };
                         }
-
                         break;
                     }
                 }
 
-                DotNet(strykerArgs.RenderForExecution(), workingDirectory: sourceProject.Path.Parent);
+                strykerArgs.SetProcessWorkingDirectory(sourceProject.Path.Parent);
+                Stryker(strykerArgs);
             }
         });
 
-    internal Configure<Arguments> StrykerArgumentsSettingsBase => _
-        => _
-           .When(IsLocalBuild, args => args.Add("--open-report:{value}", "html"))
-           .WhenNotNull(StrykerDashboardApiKey,
-                        (args, apiKey) => args.Add("--dashboard-api-key {value}", apiKey, secret: true)
-                                              .Add("--reporter dashboard"))
-           .Add("--reporter markdown")
-           .Add("--reporter html")
-           .When(IsLocalBuild, args => args.Add("--reporter progress"));
+    internal Configure<StrykerSettings> StrykerArgumentsSettingsBase => options
+        => options.When(_ => IsLocalBuild, (settings) => settings.SetOpenReport(StrykerOpenReport.Html)
+                                                                    .SetReporters([..settings.Reporters ?? [], StrykerReporter.Progress]))
+            .When(_ => StrykerDashboardApiKey is not null, settings => settings.SetDashboardApiKey(StrykerDashboardApiKey)
+                .SetReporters([..settings.Reporters ?? [], StrykerReporter.Dashboard]));
 
     /// <summary>
     /// Configures arguments that will be used by when running Stryker tool
     /// </summary>
-    Configure<Arguments> StrykerArgumentsSettings => _ => _;
+    Configure<StrykerSettings> StrykerArgumentsSettings => options => options;
 }
 
 /// <summary>
 /// Wraps information on mutation tests for a specific project.
 /// </summary>
+[UsedImplicitly]
 public record MutationProjectConfiguration
 {
     /// <summary>
@@ -252,15 +245,15 @@ public record MutationProjectConfiguration
 /// Represents the configuration for a mutation test run.
 /// </summary>
 /// <remarks>
-/// The `MutationTestRunConfiguration` class contains information such as the module name, test projects, frameworks, and project info.
+/// The <see cref="MutationTestRunConfiguration"/> class contains information such as the module name, test projects, frameworks, and project info.
 /// </remarks>
 /// <example>
 /// <code>
 /// MutationTestRunConfiguration config = new MutationTestRunConfiguration
 /// {
 ///     Module = "MyModule",
-///     TestsProjects = new string[] { "TestProject1", "TestProject2" },
-///     Frameworks = new string[] { "netcoreapp3.1", "net5.0" },
+///     TestsProjects = ["TestProject1", "TestProject2" ],
+///     Frameworks = [ "netcoreapp3.1", "net5.0" ],
 ///     ProjectInfo = new StrykerProjectInfo
 ///     {
 ///         Module = "MyProject",