+++ src/Candoumbe.Pipelines/Components/IRestore.cs
@@ -1,8 +1,10 @@
-﻿using Nuke.Common;
+﻿using System.Linq.Expressions;
+using Nuke.Common;
 using Nuke.Common.Tooling;
 using Nuke.Common.Tools.DotNet;
 
 using static Nuke.Common.Tools.DotNet.DotNetTasks;
+using static System.Linq.Expressions.ExpressionExtensions;
 
 namespace Candoumbe.Pipelines.Components;
 
@@ -23,7 +25,10 @@ public interface IRestore : INukeBuild, IHaveSolution
                 .Apply(RestoreSettings)
             );
 
-            DotNetToolRestore();
+            //TODO remove the settings => settings part once Nuke 9.0.3 is out
+            DotNetToolRestore(s => s
+                .Apply(RestoreToolSettings)
+            );
         });
 
     internal sealed Configure<DotNetRestoreSettings> RestoreSettingsBase => _ => _
@@ -36,7 +41,12 @@ public interface IRestore : INukeBuild, IHaveSolution
     Configure<DotNetRestoreSettings> RestoreSettings => _ => _;
 
     /// <summary>
-    /// Defines when set to <see langword="true"/> if unreacheable sources should make the "restore" process fail.
+    /// Options used to restore tools dependencies
+    /// </summary>
+    Configure<DotNetToolRestoreSettings> RestoreToolSettings => _ => _;
+
+    /// <summary>
+    /// Defines when set to <see langword="true"/> if unreachable sources should make the "restore" process fail.
     /// </summary>
     [Parameter("Ignore unreachable sources during " + nameof(Restore))]
     public bool IgnoreFailedSources => TryGetValue<bool?>(() => IgnoreFailedSources) ?? false;