+++ CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.13.0] / 2025-04-11
+### 🚀 New features
+
+- Added `IIntegrationTest` component that can run integration tests ([#173](https://github.com/candoumbe/Pipelines/issues/173))
+- Added `IReportUnitTestCodeCoverage` component
+- Added `IReportIntegrationTestCodeCoverage` component
+- Added `IRestore.RestoreToolSettings` to configure the behavior of `dotnet tool restore` command.
+- Added "native" stryker tool support
+
+### 🚨 Breaking changes
+- Dropped `net6.0` support
+  -Separated directories for code coverage and coverage history between unit and integration tests.
+- `IPushNuGetPackages` component no longer add `--skip-duplicate` option by default.
+- Replaced `Configure<DotNetRunSettings> StrykerArgumentSettings` is now `Configure<StrykerSettings>`
+
+### 🛠️ Fixes
+- `IBuildDockerImage` component should run AFTER `IUnitTest`, `IIntegrationTest` and `IMutationTest` components.
+
 ## [0.12.1] / 2024-11-12
 ### 🛠️ Fixes
 - Fixed incorrect verbosity of the default implementation of the `IUnitTest` component 
@@ -218,7 +236,8 @@ So now `{MutationTestDirectory}/[{framework}]` is now changed to `{MutationTestD
 ## [0.1.0] / 2022-10-23
 - Initial release
 
-[Unreleased]: https://github.com/candoumbe/Pipelines/compare/0.12.1...HEAD
+[Unreleased]: https://github.com/candoumbe/Pipelines/compare/0.13.0...HEAD
+[0.13.0]: https://github.com/candoumbe/Pipelines/compare/0.12.1...0.13.0
 [0.12.1]: https://github.com/candoumbe/Pipelines/compare/0.11.0...0.12.1
 [0.11.0]: https://github.com/candoumbe/Pipelines/compare/0.10.0...0.11.0
 [0.10.0]: https://github.com/candoumbe/Pipelines/compare/0.9.0...0.10.0